# api/swagger.json
{"openapi":"3.0.0","paths":{"/api/authors":{"post":{"operationId":"AuthorsController_create","summary":"Create an author","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorCreateCommand"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/327f7ac0f1b801893f42d"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["authors"]},"get":{"operationId":"AuthorsController_search","summary":"List and search authors","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/27f7ac0f1b801893f42df"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["authors"]}},"/api/authors/{id}":{"get":{"operationId":"AuthorsController_read","summary":"Read an author","parameters":[{"name":"id","required":true,"in":"path","example":6,"schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/7f7ac0f1b801893f42dfe"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["authors"]},"put":{"operationId":"AuthorsController_update","summary":"Update an author","parameters":[{"name":"id","required":true,"in":"path","example":6,"schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorUpdateCommand"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/f7ac0f1b801893f42dfe8"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["authors"]},"delete":{"operationId":"AuthorsController_delete","summary":"Delete an unused author","parameters":[{"name":"id","required":true,"in":"path","example":6,"schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/7ac0f1b801893f42dfe88"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["authors"]}},"/api/subjects":{"post":{"operationId":"SubjectsController_create","summary":"Create a subject","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectCreateCommand"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ac0f1b801893f42dfe881"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["subjects"]},"get":{"operationId":"SubjectsController_search","summary":"List and search subjects","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/c0f1b801893f42dfe8819"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["subjects"]}},"/api/subjects/{id}":{"get":{"operationId":"SubjectsController_read","summary":"Read an subject","parameters":[{"name":"id","required":true,"in":"path","example":6,"schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/0f1b801893f42dfe88192"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["subjects"]},"put":{"operationId":"SubjectsController_update","summary":"Update an subject","parameters":[{"name":"id","required":true,"in":"path","example":6,"schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectUpdateCommand"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/f1b801893f42dfe881929"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["subjects"]},"delete":{"operationId":"SubjectsController_delete","summary":"Delete an unused subject","parameters":[{"name":"id","required":true,"in":"path","example":6,"schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/1b801893f42dfe881929a"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["subjects"]}},"/api/books":{"post":{"operationId":"BooksController_create","summary":"Create a book","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookCreateCommand"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/b306b327f7ac0f1b80189"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["books"]},"get":{"operationId":"BooksController_search","summary":"List and search books","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/306b327f7ac0f1b801893"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["books"]}},"/api/books/{id}":{"get":{"operationId":"BooksController_read","summary":"Read a book","parameters":[{"name":"id","required":true,"in":"path","example":6,"schema":{"type":"number"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/06b327f7ac0f1b801893f"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["books"]},"put":{"operationId":"BooksController_update","summary":"Update an author","parameters":[{"name":"id","required":true,"in":"path","example":6,"schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookUpdateCommand"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/6b327f7ac0f1b801893f4"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["books"]},"delete":{"operationId":"BooksController_delete","summary":"Delete an unused author","parameters":[{"name":"id","required":true,"in":"path","example":6,"schema":{"type":"number"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/b327f7ac0f1b801893f42"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpError"}}}}},"tags":["books"]}}},"info":{"title":"Spassu Books API","description":"The Spassu Books API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthorCreateCommand":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"AuthorCreatedResponse":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","createdAt","updatedAt"]},"327f7ac0f1b801893f42d":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/AuthorCreatedResponse"}},"required":["success","data"]},"HttpError":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"code":{"type":"string","example":"author.not.found"},"timestamp":{"type":"string","example":"2024-05-10T12:57:59.129Z"},"path":{"type":"string","example":"/api/authors/6"}},"required":["statusCode","code","timestamp","path"]},"AuthorReadResponse":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","createdAt","updatedAt"]},"27f7ac0f1b801893f42df":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"type":"array","items":{"$ref":"#/components/schemas/AuthorReadResponse"}}},"required":["success","data"]},"7f7ac0f1b801893f42dfe":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/AuthorReadResponse"}},"required":["success","data"]},"AuthorUpdateCommand":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"AuthorUpdatedResponse":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","createdAt","updatedAt"]},"f7ac0f1b801893f42dfe8":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/AuthorUpdatedResponse"}},"required":["success","data"]},"AuthorDeletedResponse":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","createdAt","updatedAt"]},"7ac0f1b801893f42dfe88":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/AuthorDeletedResponse"}},"required":["success","data"]},"SubjectCreateCommand":{"type":"object","properties":{"description":{"type":"string"}},"required":["description"]},"SubjectCreatedResponse":{"type":"object","properties":{"id":{"type":"number"},"description":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","description","createdAt","updatedAt"]},"ac0f1b801893f42dfe881":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/SubjectCreatedResponse"}},"required":["success","data"]},"SubjectReadResponse":{"type":"object","properties":{"id":{"type":"number"},"description":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","description","createdAt","updatedAt"]},"c0f1b801893f42dfe8819":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"type":"array","items":{"$ref":"#/components/schemas/SubjectReadResponse"}}},"required":["success","data"]},"0f1b801893f42dfe88192":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/SubjectReadResponse"}},"required":["success","data"]},"SubjectUpdateCommand":{"type":"object","properties":{"description":{"type":"string"}},"required":["description"]},"SubjectUpdatedResponse":{"type":"object","properties":{"id":{"type":"number"},"description":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","description","createdAt","updatedAt"]},"f1b801893f42dfe881929":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/SubjectUpdatedResponse"}},"required":["success","data"]},"SubjectDeletedResponse":{"type":"object","properties":{"id":{"type":"number"},"description":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","description","createdAt","updatedAt"]},"1b801893f42dfe881929a":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/SubjectDeletedResponse"}},"required":["success","data"]},"Currency":{"type":"string","enum":["BRL","USD","EUR"]},"BookCreateCommand":{"type":"object","properties":{"title":{"type":"string"},"publisher":{"type":"string"},"edition":{"type":"number"},"year":{"type":"string","example":"2020"},"value":{"type":"number"},"currency":{"example":"EUR","$ref":"#/components/schemas/Currency"},"authors":{"type":"array","items":{"type":"string"}},"subjects":{"type":"array","items":{"type":"string"}}},"required":["title","publisher","edition","year"]},"BookCreatedResponse":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"publisher":{"type":"string"},"edition":{"type":"number"},"year":{"type":"string"},"value":{"type":"number"},"currency":{"type":"string"},"subjects":{"type":"array","items":{"$ref":"#/components/schemas/SubjectReadResponse"}},"authors":{"type":"array","items":{"$ref":"#/components/schemas/AuthorReadResponse"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","publisher","edition","year","value","currency","subjects","authors","createdAt","updatedAt"]},"b306b327f7ac0f1b80189":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/BookCreatedResponse"}},"required":["success","data"]},"BookReadResponse":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"publisher":{"type":"string"},"edition":{"type":"number"},"year":{"type":"string"},"value":{"type":"number"},"currency":{"type":"string"},"subjects":{"type":"array","items":{"$ref":"#/components/schemas/SubjectReadResponse"}},"authors":{"type":"array","items":{"$ref":"#/components/schemas/AuthorReadResponse"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","publisher","edition","year","value","currency","subjects","authors","createdAt","updatedAt"]},"306b327f7ac0f1b801893":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"type":"array","items":{"$ref":"#/components/schemas/BookReadResponse"}}},"required":["success","data"]},"06b327f7ac0f1b801893f":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/BookReadResponse"}},"required":["success","data"]},"BookUpdateCommand":{"type":"object","properties":{"title":{"type":"string"},"publisher":{"type":"string"},"edition":{"type":"number"},"year":{"type":"string","example":"2020"},"value":{"type":"number"},"currency":{"example":"EUR","$ref":"#/components/schemas/Currency"},"authors":{"type":"array","items":{"type":"string"}},"subjects":{"type":"array","items":{"type":"string"}}},"required":["title","publisher","edition","year"]},"BookUpdatedResponse":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"publisher":{"type":"string"},"edition":{"type":"number"},"year":{"type":"string"},"value":{"type":"number"},"currency":{"type":"string"},"subjects":{"type":"array","items":{"$ref":"#/components/schemas/SubjectReadResponse"}},"authors":{"type":"array","items":{"$ref":"#/components/schemas/AuthorReadResponse"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","publisher","edition","year","value","currency","subjects","authors","createdAt","updatedAt"]},"6b327f7ac0f1b801893f4":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/BookUpdatedResponse"}},"required":["success","data"]},"BookDeletedResponse":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"publisher":{"type":"string"},"edition":{"type":"number"},"year":{"type":"string"},"value":{"type":"number"},"currency":{"type":"string"},"subjects":{"type":"array","items":{"$ref":"#/components/schemas/SubjectReadResponse"}},"authors":{"type":"array","items":{"$ref":"#/components/schemas/AuthorReadResponse"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","publisher","edition","year","value","currency","subjects","authors","createdAt","updatedAt"]},"b327f7ac0f1b801893f42":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"$ref":"#/components/schemas/BookDeletedResponse"}},"required":["success","data"]}}}}